New Development: Yes/No             JIRA Issue: CAS-2594         Ready for Test: Yes/No      Interface Changes: Yes/No What Interface Changed: Please list interface changes          Test Programs: List test programs   Put in Release Notes: No              Module(s): Module Names change impacts.            Description: Fix the index computation for >32 agents
git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@13572 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ code/flagging/implement/Flagging/RFFlagCube.h
@@ -233,7 +233,7 @@ public:
     // RESET or IGNORE policy, or to corrmask for HONOR
     check_rowmask,   // same for row flags: 0 or RowFlagged
     my_corrflagmask; // see above
-  unsigned flagmask_kiss; // represents a bitmask with only bit number <n> set where 
+  unsigned long flagmask_kiss; // represents a bitmask with only bit number <n> set where 
                           // <n> is the value of this variable
   static Int agent_count;    // # of agents instantiated
   static RFlagWord base_flagmask, // flagmask of first agent instance
